requirements.txt
absl-py==1.4.0
accelerate==0.18.0
addict==2.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==0.5.2
altair==4.2.2
antlr4-python3-runtime==4.8
anyio==3.6.2
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
basicsr==1.4.2
beautifulsoup4==4.12.2
black==21.4b2
blinker==1.6.2
boto3==1.26.123
botocore==1.29.123
braceexpand==0.1.7
brotlipy==0.7.0
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.3
coloredlogs==15.0.1
contourpy==1.0.7
controlnet-aux==0.0.3
cryptography==39.0.1
cycler==0.11.0
Cython==0.29.34
dataclasses-json==0.5.7
decorator==5.1.1
decord==0.6.0
-e git+https://github.com/facebookresearch/detectron2.git@d1e04565d3bec8719335b88be9e9b961bf3ec464#egg=detectron2
diffusers==0.15.1
easydict==1.9
easyocr==1.6.2
einops==0.6.1
encodec==0.1.1
entrypoints==0.4
fairscale==0.4.3
fastapi==0.95.1
ffmpeg-python==0.2.0
ffmpy==0.3.0
filelock==3.12.0
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
ftfy==6.1.1
funcy==2.0
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==3.28.1
gradio-client==0.1.3
graphviz==0.8.4
greenlet==2.0.2
groundingdino==0.1.0
grpcio==1.54.0
h11==0.14.0
h5py==3.8.0
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.13.4
humanfriendly==10.0
hydra-core==1.1.0
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
imgaug==0.4.0
importlib-metadata==6.5.0
importlib-resources==5.12.0
invisible-watermark==0.1.5
iopath==0.1.9
jax==0.4.6
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.4
kornia==0.5.0
langchain==0.0.101
lazy-loader==0.2
libclang==16.0.0
linkify-it-py==2.0.0
lit==16.0.2
llvmlite==0.39.1
lmdb==1.4.1
lvis==0.5.3
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
mxnet-mkl==1.6.0
mypy-extensions==1.0.0
netaddr==0.8.0
networkx==3.1
ninja==1.11.1
numba==0.56.4
numpy==1.23.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.0.131
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.1.2
onnx==1.13.1
onnxruntime==1.14.1
open-clip-torch==2.17.1
openai==0.27.4
openai-whisper==20230314
opencv-python==4.7.0.72
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
orjson==3.8.10
packaging==23.1
pandas==1.5.3
pathspec==0.11.1
Pillow==9.5.0
pip-search==0.0.12
pkgutil-resolve-name==1.3.10
portalocker==2.7.0
prettytable==3.7.0
protobuf==3.20.3
psutil==5.9.5
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post4
pycocotools==2.0.2
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
pydeck==0.8.1b0
pyDeprecate==0.3.1
pydot==1.4.2
pydub==0.25.1
pyee==9.0.4
Pygments==2.15.1
Pympler==1.0.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyre-extensions==0.0.23
pyrsistent==0.19.3
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==1.5.10
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
regex==2023.3.23
requests==2.29.0
requests-oauthlib==1.3.1
rich==13.3.4
rsa==4.9
s3transfer==0.6.0
sacremoses==0.0.53
safetensors==0.3.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.9.1
segment-anything==1.0
semantic-version==2.10.0
sentencepiece==0.1.98
setuptools-rust==1.6.0
shapely==2.0.1
simplet5==0.1.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.47
starlette==0.26.1
streamlit==1.21.0
suno-bark==0.0.1a0
supervision==0.4.0
sympy==1.11.1
tabulate==0.9.0
tb-nightly==2.13.0a20230420
tenacity==8.2.2
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorrt==8.6.0
termcolor==2.2.0
test-tube==0.7.5
threadpoolctl==3.1.0
tifffile==2023.4.12
tiktoken==0.3.1
timm==0.6.13
tokenizers==0.11.4
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.0
torchaudio==0.13.0
torchmetrics==0.11.4
torchvision==0.14.0
tornado==6.3
tqdm==4.65.0
transformers==4.28.0
triton==2.0.0
typing-extensions==4.5.0
typing-inspect==0.8.0
tzdata==2023.3
tzlocal==4.3
uc-micro-py==1.0.1
urllib3==1.26.15
uvicorn==0.21.1
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
webdataset==0.2.48
WebOb==1.8.7
websockets==11.0.2
Werkzeug==2.2.3
wget==3.2
wldhx.yadisk-direct==0.0.6
wrapt==1.14.1
yacs==0.1.8
yapf==0.33.0
yarl==1.8.2
zipp==3.15.0
